+++ src/compliment/sources/keywords.clj
@@ -1,14 +1,14 @@
 (ns compliment.sources.keywords
   "Completion for keywords interned globally across the application"
   (:require [compliment.sources :refer [defsource]]
-            [compliment.utils :refer [defmemoized resolve-namespace]])
+            [compliment.utils :refer [resolve-namespace]])
   (:import java.lang.reflect.Field))
 
-(defmemoized ^:private keywords-table
-  []
-  (let [^Field field (.getDeclaredField clojure.lang.Keyword "table")]
-    (.setAccessible field true)
-    (.get field nil)))
+(def ^:private keywords-table
+  (delay
+    (let [^Field field (.getDeclaredField clojure.lang.Keyword "table")]
+      (.setAccessible field true)
+      (.get field nil))))
 
 (defn- tagged-candidate [c]
   {:candidate c, :type :keyword})
@@ -19,7 +19,7 @@
   [prefix ns]
   (let [prefix (subs prefix 2)
         ns-name (str ns)]
-    (for [[kw _] (keywords-table)
+    (for [[kw _] @keywords-table
           :when (= (namespace kw) ns-name)
           :when (.startsWith (name kw) prefix)]
       (tagged-candidate (str "::" (name kw))))))
@@ -41,7 +41,7 @@
   [prefix ns]
   (when-let [[_ alias prefix] (re-matches #"::([^/]+)/(.*)" prefix)]
     (let [alias-ns-name (str (resolve-namespace (symbol alias) ns))]
-      (for [[kw _] (keywords-table)
+      (for [[kw _] @keywords-table
             :when (= (namespace kw) alias-ns-name)
             :when (.startsWith (name kw) prefix)]
         (tagged-candidate (str "::" alias "/" (name kw)))))))
@@ -54,7 +54,7 @@
     (cond (and double-colon? has-slash?) (aliased-candidates prefix ns)
           double-colon? (concat (qualified-candidates prefix ns)
                                 (namespace-alias-candidates prefix ns))
-          single-colon? (for [[kw _] (keywords-table)
+          single-colon? (for [[kw _] @keywords-table
                               :when (.startsWith (str kw) (subs prefix 1))]
                           (tagged-candidate (str ":" kw))))))