+++ src/compliment/sources/namespaces_and_classes.clj
@@ -2,13 +2,13 @@
   "Completion for namespace and class names."
   (:require [compliment.sources :refer [defsource]]
             [compliment.utils :refer [fuzzy-matches?] :as utils]
-            [compliment.sources.class-members :refer [classname-doc]])
-  (:import java.io.File))
+            [compliment.sources.class-members :refer [classname-doc]]))
 
 (defn nscl-symbol?
   "Tests if prefix looks like a namespace or classname."
-  [x]
-  (re-matches #"[^\/\:\.][^\/\:]+" x))
+  [^String x]
+  (and (re-matches #"[^\/\:]+" x)
+       (not (= (.charAt x 0) \.))))
 
 (defn nscl-matches?
   "Tests if prefix partially matches a var name with periods as
@@ -30,9 +30,8 @@
   []
   (let [all-classes (utils/classes-on-classpath)]
     (utils/cache-last-result ::all-classes-short-names all-classes
-      (group-by #(-> (re-matches #"([^\.]+\.)*([^\.]+)" %)
-                     (nth 2))
-                (reduce into [] (vals all-classes))))))
+      (group-by (fn [^String s] (.substring s (inc (.lastIndexOf s "."))))
+                (apply concat (vals all-classes))))))
 
 (defn- analyze-import-context
   "Checks if the completion is called from ns import declaration. If so, and the
@@ -55,10 +54,9 @@
   [prefix]
   (reduce-kv (fn [l, ^String short-name, full-names]
                (if (.startsWith short-name prefix)
-                 (concat l (map (fn [c] {:candidate c, :type :class})
-                                full-names))
+                 (into l (map (fn [c] {:candidate c, :type :class})) full-names)
                  l))
-             ()
+             []
              (all-classes-short-names)))
 
 (defn- get-classes-by-package-name
@@ -69,7 +67,7 @@
                         (some #(.startsWith ^String % pkg-name) full-names))
                  (conj l {:candidate short-name, :type :class})
                  l))
-             ()
+             []
              (all-classes-short-names)))
 
 (defn candidates
@@ -78,42 +76,56 @@
   (when (nscl-symbol? prefix)
     (let [has-dot (> (.indexOf prefix ".") -1)
           import-ctx (analyze-import-context context)]
-      ((comp distinct concat)
-       (for [ns-str (concat (map (comp name ns-name) (all-ns))
-                            (map name (keys (ns-aliases ns))))
-             :when (nscl-matches? prefix ns-str)]
-         {:candidate ns-str, :type :namespace})
-       (for [class-str (imported-classes ns)
-             :when (nscl-matches? prefix class-str)]
-         {:candidate class-str, :type :class})
-       (cond (= import-ctx :root) (get-all-full-names prefix)
-             import-ctx (get-classes-by-package-name prefix import-ctx))
-       ;; Fuzziness is too slow for all classes, so just startsWith.
-       ;; Also have to do clever tricks to keep the performance high.
-       (if has-dot
-         (concat (for [[root-pkg classes] (utils/classes-on-classpath)
+      (into []
+            (comp cat (distinct))
+            [(for [ns-str (concat (map (comp name ns-name) (all-ns))
+                                  (map name (keys (ns-aliases ns))))
+                   :let [[literals prefix] (utils/split-by-leading-literals prefix)]
+                   :when (nscl-matches? prefix ns-str)]
+               {:candidate (str literals ns-str), :type :namespace})
+             (for [class-str (imported-classes ns)
+                   :when (nscl-matches? prefix class-str)]
+               {:candidate class-str, :type :class})
+             (cond (= import-ctx :root) (get-all-full-names prefix)
+                   import-ctx (get-classes-by-package-name prefix import-ctx))
+             ;; For capitalized prefixes, try to complete class FQN from a short name.
+             (when (and (Character/isUpperCase (.charAt prefix 0))
+                        (not import-ctx))
+               (get-all-full-names prefix))
+             ;; If prefix doesn't contain a period, using fuziness produces too many
+             ;; irrelevant candidates.
+             (for [{:keys [^String ns-str, ^String file]} (utils/namespaces&files-on-classpath)
+                   :let [[literals prefix] (utils/split-by-leading-literals prefix)]
+                   :when (and (re-find #"\.cljc?$" file)
+                              (if has-dot
+                                (nscl-matches? prefix ns-str)
+                                (.startsWith ns-str prefix)))]
+               {:candidate (str literals ns-str), :type :namespace, :file file})
+             ;; Fuzziness is too slow for all classes, so only startsWith. Also, if no
+             ;; period in prefix, only complete root package names to maintain good
+             ;; performance and not produce too many candidates.
+             (let [all-classes (utils/classes-on-classpath)]
+               (if (or has-dot (contains? all-classes prefix))
+                 (for [[root-pkg classes] all-classes
                        :when (.startsWith prefix root-pkg)
                        ^String cl-str classes
                        :when (.startsWith cl-str prefix)]
                    {:candidate cl-str, :type :class})
-                 (for [ns-str (utils/namespaces-on-classpath)
-                       :when (nscl-matches? prefix ns-str)]
-                   {:candidate ns-str, :type :namespace}))
-         (concat (for [[^String root-pkg _] (utils/classes-on-classpath)
+                 (for [[^String root-pkg _] all-classes
                        :when (.startsWith root-pkg prefix)]
-                   {:candidate (str root-pkg "."), :type :class})
-                 (for [^String ns-str (utils/namespaces-on-classpath)
-                       :when (.startsWith ns-str prefix)]
-                   {:candidate ns-str, :type :namespace})))))))
+                   {:candidate (str root-pkg "."), :type :class})))]))))
 
 (defn doc [ns-or-class-str curr-ns]
   (when (nscl-symbol? ns-or-class-str)
-    (if-let [ns (find-ns (symbol ns-or-class-str))]
-      (str ns "\n" (:doc (meta ns)) "\n")
-      (when-let [class (try (ns-resolve curr-ns (symbol ns-or-class-str))
-                            (catch Exception ex nil))]
-        (when (= (type class) Class)
-          (classname-doc class))))))
+    (let [strip-literals (comp second utils/split-by-leading-literals)
+          ns-or-class-sym (symbol (strip-literals ns-or-class-str))]
+      (if-let [ns (or (find-ns ns-or-class-sym)
+                      (get (ns-aliases curr-ns) ns-or-class-sym))]
+        (str ns "\n" (:doc (meta ns)) "\n")
+        (when-let [class (try (ns-resolve curr-ns ns-or-class-sym)
+                              (catch Exception ex nil))]
+          (when (= (type class) Class)
+            (classname-doc class)))))))
 
 (defsource ::namespaces-and-classes
   :candidates #'candidates